+++ src/main/java/com/flowergarden/flowers/Flower.java
@@ -2,9 +2,11 @@ package com.flowergarden.flowers;
 
 import com.flowergarden.properties.Freshness;
 
+import java.math.BigDecimal;
+
 public interface Flower<T> {
 
 	Freshness<T> getFreshness();
-	float getPrice();
+	BigDecimal getPrice();
 	int getLenght();
 }
