+++ src/main/java/com/flowergarden/flowers/Rose.java
@@ -4,19 +4,21 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 import com.flowergarden.properties.FreshnessInteger;
 
+import java.math.BigDecimal;
+
 @XmlRootElement
 public class Rose extends GeneralFlower {
 	
 	private boolean spike;
 	
-	public Rose(boolean spike, int lenght, float price, FreshnessInteger fresh){
+	public Rose(boolean spike, int lenght, BigDecimal price, FreshnessInteger fresh){
 		this.spike = spike;
 		this.lenght = lenght;
 		this.price = price;
 		this.freshness = fresh;
 	}
+
 	public Rose(){
-		
 	}
 	
 	public boolean getSpike(){
